[concat] Add centering options
Adds the `--center-vertically` and `--center-horizontally` to the
script. Out of the two, `--center-horizontally` is the more useful one,
since `--center-vertically` tends to lead to empty space around a row,
wihch doesn't look too good, but the option is provided nonetheless.

These two options tell the `convert` command to center images vertically
when forming a row by horizontally concatenating images, or horizontally
when vertically concatenating rows